feat(6853): extract task operations modules
