+++ crates/kamn-core/src/task_operations/models.rs
@@ -1,0 +1,9 @@
+mod domain;
+mod error;
+
+pub use domain::{
+    SwarmTaskDraft, TaskOperationNoticeKind, TaskOperationRecord, TaskOperationRecordSnapshot,
+    TaskOperationSnapshot, TASK_OPERATION_SNAPSHOT_SCHEMA_VERSION,
+};
+pub use error::{TaskOperationError, TaskOperationSnapshotStoreError};
+pub(crate) use error::lifecycle_error;
