+++ crates/kamn-core/src/task_operations/tests.rs
@@ -1,0 +1,3 @@
+mod engine_contract_tests;
+mod file_store_contract_tests;
+mod support;
